+++ medLayer/base/__init__.py
@@ -1,1 +1,1 @@
-from .node import Node
+from .event import Event
