+++ medLayer/base/interactions/coexisting_relation.py
@@ -1,5 +1,5 @@
 from medLayer.core.datatype.probability import Probability
-from medLayer.interface.irelation import IRelation
+from medLayer.core.interface.irelation import IRelation
 
 
 class CoexistingRelation(IRelation):
